backend/requirements/base.txt
Django==4.0.2
djangorestframework==3.13.1
psycopg2-binary==2.9.3
django-cors-headers==3.11.0